docker added
- repo refactored, \n - services standartized \n - config updated \n - auth completed

+++ internal/mq/serial/credit_transaction.go
@@ -2,7 +2,7 @@ package serial
 
 // RabbitMQ serializer
 type TransactionInfo struct {
-	SenderID         int `json:"sender_id"`
-	RecipientID      int `json:"recipient_id"`
-	AmountToTransfer int `json:"amountToTransfer"`
+	SenderID         uint64 `json:"sender_id"`
+	RecipientID      uint64 `json:"recipient_id"`
+	AmountToTransfer int    `json:"amountToTransfer"`
 }
